+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/mapper/HandlerMapping.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe;
+package webmvc.org.springframework.web.servlet.mvc.tobe.mapper;
 
 import jakarta.servlet.http.HttpServletRequest;
